Balance edits
Growth rates in DeCiv are already insane once you escape the early-game; being able to build a Factory Farm in all of your cities in the mid-game is too much to add on top of that. As for Mobile Encampment, the "move own capital" gimmick borrowed from Lekmod that would make sense for New Star triggers a bug where you can raze the second capital and bring it down to 0 pop without destroying the city.