Some ruins rebalancing
The extra Pantheon effect was bugged insofar as triggering it does not provide a notification. That and a test game where it triggered *three* times suggests it really shouldn't be present, at least not until further rebalancing to reduce the frequency of high power ruins bonuses.
In exchange, a stronger Faith award bonus is added and the Wandering Doctor can show up again.