Add Tidal Array + Deseret UB adjustments
There will certainly be more Power consuming buildings added to this mod in the future, and the Tidal Array is there to offset this. Yes, coastal cities already have huge advantages over non-coastal cities, but this was the most feasible option at the time for an additional Power generating building.
The Deseret Data Center should be the last UB granted to Deseret Nation, though I am iffy on this one existing. It exists purely because of the existence of the Utah Data Center IRL. Likewise, the Blue Castle Reactor added in a previous update is based on the Blue Castle Project. These are the only UBs for Deseret that are disconnected from the religious theme.
The Restored Temple of Zion gains a Culture boost after discovering Flight to assist in achieving the intended Culture wincon, in the event that said wincon actually becomes viable.
However, there must be a price to pay for these new boosts. Negative effects have been added to the Bishop's Storehouse - possessing specific luxuries will result in a Production and Faith penalty in all of your cities. Drugs are bad, m'kay?