Restore the Artificial atoll (3/3) + balance adjustments
With Artificial atolls restored, God of the Sea's buff to provide happiness for each of them is also restored.
Faith Healers is now a Follower effect, has some of its added healing moved into a new special promotion, and generates more Faith.
Primacy's old effect was merged into Religious Unity and it is given a new effect that will be of greater benefit to theocratic civs.
Art of War no longer expires in the Industrial Era; however, the Faith cost of purchase has been set at a fixed 200. This is a heavy nerf to the belief in the early-game in exchange for making it useful in the late-game.
Labor of Love is nerfed: the production bonus caps at 10% instead of 15%.
New Zion is nerfed: it no longer adds Science and Happiness to Great Improvements.
Missionary Zeal is buffed to grant Philosopher units an extra charge.
Religious Troubadors is buffed from +2 Faith per Trade Route to +3, in the hopes that this will be enough to make the AI select it for their religion.