Revert some previous changes
After repeated game crashes, I suspect that there is a serious problem with the AI routines not knowing how to handle what happens with the effect of [Papal] Primacy is changed to something other than a diplomatic bonus, causing fatal errors each time they attempt to found a religion only for the belief they're programmed to take first no longer existing. Thus I was forced to undo the Primacy + Religious Unity changes. On an unrelated note, Trust the Science is nerfed to remove the Science/Faith bonus to Academies.