New oceanic TI + Mariners support + other edits
Coastal cities become even stronger in the late-game with the Floating Settlement TI. However, the Artificial atoll is nerfed to remove the Culture yield (which may be awarded to a new building in the future). Tidal Array loses the +1 to Production on Water resources, but provides 2 additional Power to make it worth spending precious Aluminum to build.
The Mariners receive their first new support for their part of the Domination Civ support update focus.
Deseret Nation just gets flavor text updates.
The Defiant Few are now an even greater game-breaking CS when they spawn as Von Bach Industries is buffed to award them with a free Lawkeeper. (Might undo this later and move this free unit to a new UB.)